fix: Guard against variable reuse in inference
As we no longer want to use a global, persistent `Fresher` we run the
risk of reusing the same variable. For imports we can just freshen the
type as they are used. For environments we do not need to worry about
read-only environments as no substitution will be applied to them
however for environments that can be written to we now check that the
environment does not contain any variables that the fresher may emit.

If `PolyType`s took variable shadowing into account during substitution
then we shouldn't need to worry about this but I don't think the code is
in a place where this is possible just yet.