fix: reject usePrevious when column doesn't exist
This patch addresses an issue where one might do
`fill(column: "a-nonexistent-column" usePrevious: true)`. Why would
someone do that, you might ask. In the current bug being chased, it was
because the table was a result of a `union` of two different schemas,
and sometimes the fill happened when the other schema was the first
record in the `union`.