feat(zoneinfo): add methods to convert to and from local clock time
This adds methods to the `Location` struct to convert from local clock
time to a UTC version of the same clock time and to convert to the local
clock time from the UTC version of the same clock time.

This will be used by the time zone support to translate a local time to
a version that we can perform time math on and then to convert back from
the time math to the local clock time.