chore: Avoid the need to run make generate after modifying the rust code
Removes `buildinfo.gen.go` from git so that changes to it do not cause
merge conflicts however it still exists and are generated by `make
generate` so changes in it will cause cgo to run again. The issue still
remaining is to assert that the go code does not use a cached version of
the rust library so I embedded the git revision into the rust library
and added a test to check that it matches to the one seen when running
the go tests. Since `buildinfo.gen.go` will be generated fresh on each
CI call that forces cgo to use the correct version of the rust build.