feat(interval): add location functionality to interval package
This adds the location functionality to the interval package using the
`zoneinfo.Location` struct from the `zoneinfo` package.

The interval package now translates an input time from the local clock
time to the utc clock time if a location has been specified. It performs
the same time math it previously did and then it will convert the times
back to the local clock time before returning them.

If the default `NewWindow` is used or the location is nil, it continues
to use the old behavior which does not translate for zone offsets.