feat(stdlib): expose location functionality to the window transformation
This exposes the location functionality to the window and
aggregateWindow functions. The location can be specified as an optional
parameter to customize the location that the windowing happens in.

By default, the windowing happens in the location specified by the
`location` option defined in package `universe`. The default value for
this is `utc` which is the equivalent of the string `"UTC"` and disables
location functionality entirely.

This also modifies the `window()` transformation to accept `0s` as an
argument to `window()`. When `0s` is used as a parameter to `every` and
`period`, it tells window to use the default value. The default value
for `every` is `period` and the default value for `period` is `every`.
If both are zero, an error is returned.

The `aggregateWindow()` function now also exposes the `period` parameter.