feat: add timezone package with fixed offset location
The timezone package now exists for loading locations and fixed offsets.
The `location` function inside of `timezone` will verify the location
exists so that it doesn't have to be done by the `window()` function.

The `fixed` option creates a location in the `UTC` zone with an
additional offset that is separate from the normal offset. This offset
ends up working in reverse of the typical offset. An option of `-4h`
will set the offset to be at `-04:00` permanently as if it were in UTC.
The practical effect of this is that these two are equivalent:

    window(..., offset: 4h)
    window(..., location: timezone.fixed(offset: -4h))