refactor: change the plan window spec to use a basic data struct instead of the location struct
This changes the `plan.WindowSpec` to use a basic data struct for the
location definition instead of loading the location and including the
struct.

Loading the location was convenient for flux, but it was less convenient
for planner rules which needed to interact with basic data types. This
specifically caused a problem with window related pushdowns. While we
could check that the location was proper because of exposed types, it
relied on Go comparison mechanisms that made me uncomfortable.

This causes the loading of a location to again be moved into the
creation of the transformation. This shouldn't cause any kind of real
change because we already validate the loaded location in the timezone
package.