refactor(interval): modify interval to use nil as the zone for utc
The original design of interval checked for a nil value and ignored the
conversion to/from the local clock when creating windows. When we added
fixed offsets, the location was changed to a real UTC location. This
invalidated the old code so UTC times would go through the local clock
even though that would cause an unnecessary performance hit because it
would go through the `FromLocalClock` code.

This changes the interval code to have UTC load a nil location so that
it skips this code again. This makes it more compatible with the default
value and makes comparisons on the pointer value easier.