chore(libflux): refactor to use anyhow::Error (#4131)
Prior to this change the libflux code has an inconsistent use of String
and custom Error structs that just wrapped strings to represent errors.
Now with this change those instances have been replaced with
anyhow::Error as a simple error type for when we needed a plain error.

We still have 4 uses of custom error structs/enums where we have very
specific information we want to capture on the error. Perhaps we should
update those to use https://github.com/dtolnay/thiserror a companion to
anyhow in the future.