perf(execute): Avoid recompiling types for every Process call
When executing `Process` on a table with few rows the time to `Compile`
the function before execution ends up dominating the total execution
time. This is an issue for data produced by `csv.from` which creates one
`Table` per row and might be for other sources as well though I have no
idea how common it would be.

For running the following query on a file with 50k lines this change
reduces the runtime from ~38s to ~16s.

```
import "csv"

csv.from(file: "../indata_dup.csv")
    |> filter(fn: (r) => r.Animal_name == "crow")
    |> count()
```

To make `compileFn` immutable after compilation (at least as far as I
and `make test` understands) I changed it to only hold the parent scope
and only construct the scope used during evaluation when `Eval` is
actually called.