feat(execute): add disposable interface for transformations
This adds a `Disposable` interface to each of the new transformation
types and implements the usage of that interface.

The `Dispose` method signals that a resource is unused and can be
disposed of. An implementation of this can also use the method to signal
that it should be reused.

All transformations now require this method to be implemented.

The `Dispose` method is also utilized by stateful transformations to
dispose of the state if the interface is implemented. This includes both
the aggregate and narrow state transformations.