feat: Report multiple errors from type inference
Having multiple type errors in a script is fairly common so letting
users see and fix all of them at once without needing to
fix,compile,fix,compile... multiple times is more user friendly.

We could extend this further as well such that we try to typecheck an
AST with errors in it as well.