refactor: Pass a persistent Substitution throughout inference
This replaces the `Fresher` type for most use cases as the
`Substitution` becomes responsible for creating fresh type variables
(the `Fresher` is still used in the `Fresh` trait as there is no need
for a full `Substitution` there).

By having a persistent `Substitution` we will no longer need to re-check
any constraints as the `Substitution` now records it for later lookup.
Since we do not need to re-check the constraints we avoid a lot of
duplicate work and it simultaneously enables easy reporting of multiple
errors (if we re-check constraints we could get duplicate errors).