refactor: Prepare type inference for a persistent Substitution (#4151)
* refactor: Remove PartialEq from Substitution

`ena`s unification table does not implement `PartialEq`

* refactor: Introduce a struct to hold state in inference

Currently `Environment` and `Fresher` are passed as separate arguments
which makes it more work than necessary to add in more state to this
module. By adding a struct to group the state together we may avoid
this.

* refactor: Hold the `Environment` in the `Infer` struct

Removes one parameter from all `infer` functions. Passing `Environment`
by value and returning it is practically identical to passing it by
`&mut` but is just a bit less verbose.

* refactor: Introduce Substituter trait

With the persistent substitution we will not be able to create temporary
substitutions as cheaply so without can't be implemented in a good way.
This changes `apply` functions to take a trait instead to provide a way
to alter the behaviour of it instead.

* chore: make generate

* refactor: nits