feat(fluxdoc): add support for example parsing
This is the first part of a two part change to parse examples.
With this change parsing was separated from constructing the *Doc
structures. This made adding the parsing of examples much easier than in
the previous design.

The second part will include work to execute and test the examples
incorporating their inputs and outputs.