feat(fluxdoc): add executable exmaples from Flux docs (#4172)
* feat(fluxdoc): add support for example parsing

This is the first part of a two part change to parse examples.
With this change parsing was separated from constructing the *Doc
structures. This made adding the parsing of examples much easier than in
the previous design.

The second part will include work to execute and test the examples
incorporating their inputs and outputs.

* feat(fluxdoc): generate input/outputs from examples in docs

With this change the `fluxdocs` make target will now generate input and
output tables, as markdown, for examples in the docs.