chore: factor testcase transformation out of edit package
The testcase code is the only part of the edit package that depends on
the Flux parser. We'd like to be able to import it without depending on
the parser, so split it out into its own `testcase` package.

To maintain full backward API compatibility, we move all but the
`TestcaseTransform` code into a new `editlite` package and redefine
the `edit` package as a lightweight forwarder on top of `editlite` and
`testcase`.

If it's OK to break backward compatibility for this (for example because
we're pretty sure that no external users are using the `TestcaseTransform`
function, we could just change the existing `edit` package instead.