fix(dependencies/influxdb): handle http errors from influxdb writer (#4192)
This handles http errors from the influxdb writer properly. It now
recognizes the http error type from the influxdb client library and
transforms it to an equivalent flux error using the same mechanism that
the read side uses.

It also has been expanded to handle the "too many requests" code that is
used for rate limiting and to categorize these as the equivalent code in
flux which is resource exhausted. This code gets translated back to
"invalid" when flux is run within influxdb.