feat: Always continue inference on errors (#4202)
* fix: Remove the function parameter hack

With a persistent substitution this is no longer needed.

* feat: Continue inference for more error types

* feat: Avoid errors in type inference causing further errors

When we continue running type inference after an error has occurred we
need to be careful not to cause additional errors that can obscure the
actual, original error. The error type serves as a placeholder that will
not cause another error when it is encountered in later checks.