fix(fmt): Formatter formats invalid syntax into even worse invalid syntax (#4203)
Issue - Formatter was formatting an invalid syntax script.

Fix - The fix is to check AST nodes(call ast.check()) and return if any error is found to the user, before proceeding to format the script.

while fixing it we found that the AST walk doesn't actually traverse BuiltinStmt. So any errors that occur when parsing the type of builtin are just ignored. So to fix it AST walk for the BuiltinStmt has been added.

fixes: influxdata/vsflux#242

Co-authored-by: Markus Westerlind <mwesterlind@influxdata.com>