fix: Reduce the amount of extra parse errors generated
Saw some room for improvement due to #4231. Due to `fn expect` consuming
tokens until it found the correct one it is really easy for the parser
to end up trying to match the rest of the source code, producing an
error for each token, obfuscating the actual cause for the error. It is
possible that this will cause worse errors in other instances (where we
would eventually find the token we expected) but there are not such
tests atm at least.

If we want to improve the error reporting further I would suggest either
using [LALRPOP](https://github.com/lalrpop/lalrpop) and its error
reporting or try to mold this hand-written parser to something closer to
[rust-analyzer](https://github.com/rust-analyzer/rust-analyzer/tree/master/crates/parser)
.

https://rust-analyzer.github.io/blog/2020/09/16/challeging-LR-parsing.html