feat: Add is_type to query the runtime type (#4178)
* feat: Add is_type to query the runtime type

Proposal for #2159. Only capable of answering true for primitive types.
Other proposals include `isFloat` etc but I went with this as it was
quicker and is easily convertible to that approach.

* test: Add a testcase for isType

* docs: Update comment in isType

* chore: Fix the added test file

* chore: Try to fix the isType test

* chore: Allow make generate to use the docker container

* feat: Add a the Kind::Primitive

This lets us restrict `isType` to only be callable with types that it
can actually answer true for (can't call it with functions, arrays etc)

* feat: Change the Primitive kind to Basic

* test: Add isType tests in flux

* chore(cmd/test): Print the failures so that one does not need to scroll up

* fix: Correct flatc instructions for go

For the go flatbuffers files we invoke `flatc` via `go generate` so the FLATC variable does not work. Instead, factor out the installation used in docker to a bash file and install via that.

* chore: Allow formatting to run in parallel via make -j

* refactor: Move isType to the new types module

* chore: make generate