feat(compiler): introduce refactored compiler package
This introduces a refactored compiler package to `internal/compiler`.
The refactored compiler is designed to take full advantage of the
information from the type system to create an optimized pathway for
runtime execution.

Some important features of this new compiler:
- Preallocates registers for instructions.
- Executes instructions from a main loop instead of recursive descent.

The compiler compiles a function expression into a set of instructions
that operate on an infinite register VM in the mold of an SSA
instruction set. The registers are preallocated with memory for the
given type. Access to identifiers is done through register indices
instead of by name as these are computed at compilation time rather than
at runtime. Similarly, record indices are computed at compile time
instead of determined at runtime.