feat(stdlib): new parameters added to difference function to support increase function
BREAKING CHANGE:

A new boolean parameter initialZero: true has been added to the increase function with functionality as mentioned below -

1) (existing behavior)
The first row is never preserved.

(new behavior)
The first row will be kept and have a value of zero.

2) Subtraction rules for numeric types.

The difference between two non-null values is their algebraic difference; (same for existing and new)

(existing behavior)
OR null, if the result is negative and nonNegative: true

(new behavior)
OR current value, if the result is negative and the current value is greater than equal to zero and nonNegative: true
OR null, if the result is negative and the current value is less than zero and nonNegative: true

fixes: #4098