fix: add targets to rust-toolchain (#4292)
As we're using `rust-toolchain.toml` to manage our toolchain, there's
some implicit changes that come along with it. In the case of cloud, the
build system is set up for rust 1.55, but when `rustup` sees the
toolchain file, it opts to use `1.56`, which doesn't have the components
needed in it. This patch specifically adds the targets for which we need
to fetch the stdlib.