refactor: Split up write tests and read tests, run writes then reads
Two birds/one stone. We've got problems with the `sql.to()` calls
incorrectly running twice when they are sequenced ahead of an
`sql.from()` call with `experimental.chain()`.

Additionally, some db drivers are more optimistic than others and will
"hang up" on the client while the data is still not quite available on a
subsequent read. This means the tests end up running flaky for some
engines (MySQL was the first culprit in this).

By splitting up the writes and reads, we can optionally include a
`sleep` in the `Makefile` between these two steps if we end up seeing
increased flaky test runs in the future.

See: <https://github.com/influxdata/flux/issues/4302>