refactor: Solve constraints immediately (#4303)
* refactor: Solve constraints immediately

I don't think there is any particular reason for packaging up the constraints and returning them before solving them(?). If we solve them immediately we get the type information added to `Substitution` immediately, potentially allowing for better error messages if we utilize it (could help in #4277). We can also pass in references instead of cloning the values.

* chore: cargo clippy

* chore: make generate