fix: ensure arrays are not table streams before calling `Len()`
`TableObject` currently implements the `Array` interface but all the
methods panic, explaining they are unsupported.

This change aims to avoid panicking and instead give an error message
suited to each case.

A new interface is introduced, `values.ITableObject,` which allows type
assertions to be made without importing `flux` (to get access to
`flux.TableObject`). This is a tactic to avoid cyclical imports. For
cases where the module already had access to `flux.TableObject`, it's
used directly. For most other cases, `values` was already imported
making the interface available. For other cases, `values.ITableObject`
is favored to avoid more contention on `flux`.

Earlier patches targeted specific callsites that appeared in logs:

- <https://github.com/influxdata/flux/pull/4272>
- <https://github.com/influxdata/flux/pull/4276>