test: Add tests for many `TableObject` type assertions (but not all)
A handful of the new branches aren't being hit by tests. Some gaps are due
to me not understanding how to build `TableObject` values suited to the given
test (empty `TableObject`s are technically _invalid_ when it comes to
sniffing their type/nature).

Other gaps are due to the code requiring the type assertion being a
side-effect, looking to write to some remote service.