feat: add timeable support to (add/sub)Duration (#4341)
* feat: add timeable support to (add/sub)Duration

Prior to this change users had to workaround the fact that a timeable
was difficult to manipulate. For example v.timeRangeStart from the UI.
See
https://chrispnerd.blogspot.com/2021/12/i-have-application-that-feeds-periodic.html
for a good explanation.

Now users can use addDuraiton and subDuration directly.

* fix: make date aware of mixed duration types

Prior to this change the date functions timeable behavior did not
correctly honor durations that span a month. This has been fixed and
tests added. Additionally this logic was encapsulated into a single
method.