refactor(execute): refactor the disposable interface to closer
We will need the ability to return an error from places where dispose
would have been called so renaming this to `Closer` to match `io.Closer`
and changing the signature to return an error.

This does not use `io.Closer` to differentiate the interface from the
one in the standard library, but it is identical.