feat: Use reference equality for `Symbol`
Rather than using equality by name we can compare `Symbol` by the reference that it points which lets us immediately distinguish between the symbols of two bindings.

```
(x) => {
    // These two `x` below are not the same as the function parameter due shadowing
    x = 1
    return x
}
```

This gives us two benefits, first any lookups done with `Symbol` keys are faster as it only needs to compare/hash a single pointer instead of the full name. Second, in any analysis done after `convert` we don't need to care about scope to the same extent as two symbols comparing equal guarantees that they refer to the same thing which would simplify getting all references to a variable in the LSP (for instance).

This alters the flatbuffer serialization and deserialization to use (and rely on) shared strings so that this reference equality is preserved across flatbuffers at least on the rust side, the go side does not appear to need this at the moment.