refactor: include quote-reliant column name in DDL
Expand the surface area of the integration tests to include a column
name that has a space in it, requiring it to be quoted.

This change is "double duty" in so much as it also introduces a nullable
column into the mix. This complicates the tables we assert against since
it forces the use of `union` and `debug.opaque` to be able to represent
the table we expect to get back from the database.