fix(stdlib/http): IP validation failures should be classified as Invalid
The errors returned by the HTTP Client will have a few extra layers
around them when the initial error comes from its control function.

Since the control function is where we run our IP validation, the error
code information was being "lost" at the point where it would be reported.

This change attempts to unwrap the outer layers of the Error so we can
preserve the initial cause. If the nesting is not the expected
structure, return the root error as we did before.