feat(stdlib): create an optimized sort limit
The optimized sort limit will take the pattern `sort |> limit` when
limit has no offset and will replace it with an optimized version that
performs the sorting continuously as data comes in rather than
performing a sort on all of the data and then doing the limit.

This is most analogous to performing a merge sort on each incoming
chunk of data.