fix(stdlib/sql): quote db identifiers (#4328)
* fix(stdlib/sql): quote/escape table and column identifiers

* refactor: update hdb to use quoted column identifiers

This diff closes the opening for SQL injection via column identifiers in
hdb, but highlights the inconsistency between the quote/escape strategy
used for table names and column names. Table names were already being
quoted/escaped, but were also being transformed to uppercase.

It could be argued we should do the same transformation with column
identifiers, but I would advocate _the other way_ (to not transform the
case at all anywhere).

The issue is automatically forcing identifiers to uppercase, then
quoting them, means tables/columns created via DDL in the db engine
effectively become unreachable from Flux when they:

- happen to be quoted (preserving case)
- happen to _not be uppercase_

By adding a perceived convenience with the transformation, we reduce
control and possibly make valid identifiers impossible to
reference.

For the original discussion on the strategy, see:
<https://github.com/influxdata/flux/pull/3098>

* fix: resolve issue with identifier casing in hdb

The HDB code path will try to force identifiers to all be UPPER CASE.
The proc for automatically creating the target table (when missing) only
uppercased the table name in certain situations (when specified with a
schema name stem). In the other case, the table name was incorrectly
left as-is.

HDB also handled quoting/escaping column names as a part of the
"translate func" for this driver, so the initial work to handle this
closer to the top of the `to` impl resulted in "stacked quoting" which
caused problems in SQL generation.

If we move all column escaping/quoting to happen during translation, we
can unify things a bit.

* refactor: quote idents closer to where they are templated

The column translate func now escapes column names for DDL, quoteIdent for
insert. Feels silly to do this in two places.

If the translate func returned pairs of (escaped) column name + type, then
we could use the translate func to generate identifiers for both cases.

Currently it returns a string containing both together.

* docs: add remark on ident quoting/escaping being tranlate func's job

* refactor: include quote-reliant column name in DDL

Expand the surface area of the integration tests to include a column
name that has a space in it, requiring it to be quoted.

This change is "double duty" in so much as it also introduces a nullable
column into the mix. This complicates the tables we assert against since
it forces the use of `union` and `debug.opaque` to be able to represent
the table we expect to get back from the database.

* test: update column translate tests to assert quoted DDL

The column translate functions should now be returning quoted column
identifiers.

* fix: quote/escape string literals in SQL DDL

Mitigate risk of SQL injection by escaping interior single quotes when
generating string literals for SQL DDL.

* fix: quote/escape another string literal, add comments

* refactor: mark mysql and postgres quote helpers as private

* chore: favor docker stop instead of docker rm for automatic volume cleanup

* fix: string literals escape ' as ''

* fix: identifiers escape " as ""

* test: add sql injection attempt tests

* fix: escape ` as ``, fixup hdb escaping

* test: comment out vertica injection attempt, possible driver bug

* chore: make fmt

* refactor: tighten up test code with custom assertion helpers

Many of the assertions in the `sql.to` tests contain duplicate blocks
which were originally written in a very verbose style.

Slim the tests down by pulling the repetion out into helper functions.

* refactor: hoist "seed want" values up to top of acceptance tests

* chore: make generate