fix: Remove duplicate die builtin
This is the cause for the lack of build reproducibility. Since these two definitions now gets
assigned different symbols they both end up stored in `Environment` and when we later construct
`PackageExports` the iterator returns both (before only the latter would be returned). So despite
the resulting iterator being sorted we end up with these two bindings in an arbitrary order and only on ends up in the `String` keyed map.

Since all top level bindings get exported, we may want to error on duplicates to avoid confusion about which gets exported.