fix(universe): return `false` if `contains` gets invalid value
Flux treats the access of a missing field in the `value` parameter for
`contains` as a hard error, whereas similar access in `filter` is
treated as "falsy."

As noted in <https://github.com/influxdata/influxdb/issues/23025>, it's
possible the error behavior was introduced as a breaking change accidentally
in <https://github.com/influxdata/flux/pull/2696> since a user raised the
issue as something that worked previously but  no longer does.

This patch checks to see if `value` is invalid and immediately returns
`false` if it is, sidestepping the type check between the `value` and
`set` parameters that produces the error.

Refs:
- <https://github.com/influxdata/flux/issues/4381>
- <https://github.com/influxdata/influxdb/issues/23025>
- <https://github.com/influxdata/flux/pull/2696>