refactor: skip invalid values in `set` array
If we want to be permissive when `contains` receives `Invalid` values at
runtime, it makes sense to skip any `Invalid` items in `set` in addition
to the early (`false`) return.

This comes at the cost of being able to fail fast with a strict
comparison to decide the type of the `set` array. Now we check the type
and nature of each item as we iterate.