feat(dependency): add dependency hooks to the dependency subsystem (#4421)
This adds hooks to the dependency subsystem to run certain actions at
certain times of the execution. At present, the only hook is the
`OnFinish` hook which runs when execution has finished.

The mqtt dependency has been modified to use a pool dialer by default.
The pool dialer will retrieve an `mqtt.Client` that has the same options
and brokers that was previously "closed" to avoid establishing a new
connection. This means repeated calls to `mqtt.publish()` will use the
same client.

Pooled clients are cleaned up when `OnFinish()` is called.