fix(dependencies): revert update to the dependencies package (#4433)
This reverts the change to the dependencies package. The change switched
the package to return a more generic interface rather than the struct
that held all of the dependencies, but there was existing code that used
this struct that was not easily refactorable.

The original version of the PR that introduced this change needed the
change to simplify a large change, but we ended up doing a different
change instead so this reverts the small section of the change back to
its original form just to simplify the area where this function was
consumed.