chore: disable MergeFiltersRule
The rule is _only enabled_ here in the CLI, making the behavior
inconsistent with other envs.

The rule is specifically problematic because it can cause some queries
to fail even when written explicitly to avoid type check failures.

The rule could be re-enabled at a later date if we can figure out a way
to get the type checking to be permissive enough in this case.

Refs: #4378
Fixes: #4436