fix(feature): int feature flags work properly when returned as floats (#4441)
The int type for feature flags has been switched from int32 to int since
it is more consistent. Additionally, if a float value is returned and is
meant to be interpreted as an int, the conversion now happens. This can
sometimes happen because feature flag libraries may transfer the data as
JSON and JSON doesn't have an integer type instead treating them as
floats.