feat(requests): add peek method to quickly return HTTP response
This change adds a new peek method to the requests package to make
viewing the results of an HTTP request trivial.

Additionally the examples are updated to actually consume the response of
the HTTP requests and show how to use array.from to construct tables.

The examples use real publicly and freely available APIs and can be
executed. As a result the examples are executed as part of building the
Flux docs and the results are captured.