feat(array): add filter to array functions
This change adds a filter function to the array package.
Additionally the package is moved to the experimental package
so that users can use both the new functions as well as array.from from
the same package import. Eventually these functions will also be
promoted into the normal array package.

Finally pipe arguments are added to these array functions so they can be
chained together like tables can be.