fix(schema): add stop arg to schema functions
Previously the `stop` argument could not be added to the schema
functions because the default value would overly constrain the arguments
type. That has been fixed.

Note sure how to add tests for this change as we don't actually have
tests for calling the functions rather just for the push downs matching.