feat: add display function to make representing any value simple (#4460)
Prior to this change it was not possible to convert every Flux value
into a string for easy debugging. With this change a `display` function
has been added that converts any Flux value into a string
representation.

The display function uses the same code that the REPL uses to display
the results of executing a line.

The bytes type display was changed to be a hexidecimal string instead of
a list of decimal values.