chore(release): add prep-release script
The Flux release process will now require a preparation step.
This still will make and commit any changes to the source code, i.e.
updating docs to mark the correct version a function was introduced.

The process will be:

1. Run ./prep-release.sh in Flux repo
2. Merge created PR
3. Run IDPE release script (which calls the Flux release script)

The Flux release script checks that the prep step was completed so it
can't be easily forgotten.