fix(test): remove `os.Exit` calls, allow defer `executor.Close` to run
Today, `TestExecutor`s that depend on being able to run code on teardown by
implementing `Close()` are unable to since `os.Exit()` will terminate the
process before any deferred functions are able to run.

The exiting, if deemed necessary, should probably be the responsibility
of the actual binary rather than deep in the test runner.

While it would be preferred to simply hand errors back to the caller,
this diff avoids updating the signatures of public functions by
converting the previous `Exit` calls into `panic`s. The benefit of this
is a panic will still allow deferred functions to be called as the stack
unwinds.

In the future, we may elect to clean this up further by refactoring public
functions like `Finish()` to return errors.

refactor: `TestRunner.Finish()` now returns an error

chore: update Finish docstring to match new signature

refactor: Exit(1) in Run command callback on failure