feat(planner): add second pass to physical planner for parallelization rules
Adds a second pass to the physical planner that is dedicated to parallelization
rules.

Parallelization is something we want to embark on only after all pushdowns have
succeed. Adding a second pass ensures that we can never parallelize a node and
block a pushdown.

Suppose we include parallelization rules the set of physical rules. Consider:

    range -> filter -> agg

And the rules parallelize filter:

    parallel-range -> parallel-filter -> merge -> agg

But fail to parallelize the aggregate as well, then we have blocked the
range-filter-agg pushdown. Using a second pass frees us from the challenging
task of ensuring we never block a pushdown.